+++ docs/CHANGELOG.rst
@@ -4,6 +4,10 @@
 Changelog
 =========
  
+python_openzwave 0.4.1.x:
+ * Update the default flavor : try to use a precompiled library. If can't find one, use the embeded package.
+ 
+
 python_openzwave 0.4.0.x:
 
  * New installation process via pip